CollectTypes rework
